test(mock): renamed the mock repository and moved to a shared directory, adjusted imports
