--- apps/backend/src/common/mocks/user.repository.fake.ts
@@ -1,4 +1,0 @@
-import { User } from '../../user/entities/user.entity';
-import { GenericRepositoryFake } from './generic.repository.fake'
-
-export class UserRepositoryFake extends GenericRepositoryFake<User> { }
