+++ apps/backend/src/goal/test/mocks/goal.repository.mock.ts
@@ -1,0 +1,7 @@
+import { GenericRepositoryMock } from '../../../common/mocks/generic.repository.mock';
+import { Goal } from '../../entities/goal.entity';
+
+/**
+ * A mock goal repository for testing purposes
+ */
+export class GoalRepositoryMock extends GenericRepositoryMock<Goal>{}
