+++ apps/backend/src/user/test/mocks/user.repository.mock.ts
@@ -1,0 +1,7 @@
+import { User } from '../../entities/user.entity';
+import { GenericRepositoryMock } from '../../../common/mocks/generic.repository.mock'
+
+/**
+ * A mock user repository for testing purposes
+ */
+export class UserRepositoryMock extends GenericRepositoryMock<User> { }
